requirements.txt
Django==1.8.2
dj-database-url==0.3.0
dj-static==0.0.6
django-appconf==1.0.1
django-compressor==1.5
djangorestframework==3.1.2
drf-nested-routers==0.9.0
gunicorn==19.1.1
psycopg2==2.6
six==1.8.0
static3==0.5.1
wsgiref==0.1.2
python-dotenv==0.1.2
celery==3.1.18
django-celery
